+++ src/components/BetweenGallery.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Swiper, SwiperSlide } from 'swiper/react';
 import 'swiper/css';
-import "../../css/home.css";
+import "../css/home.css";
 
 const cardData = [
   {
