+++ src/components/BgCarousel.tsx
@@ -1,8 +1,8 @@
 import { Swiper, SwiperSlide } from "swiper/react";
-import bg1 from '../../assets/img/bg_carousel/bg1.jpg';
-import bg2 from '../../assets/img/bg_carousel/bg2.jpg';
-import bg3 from '../../assets/img/bg_carousel/bg3.jpg';
-import bg4 from '../../assets/img/bg_carousel/bg4.jpg';
+import bg1 from '../assets/img/bg_carousel/bg1.jpg';
+import bg2 from '../assets/img/bg_carousel/bg2.jpg';
+import bg3 from '../assets/img/bg_carousel/bg3.jpg';
+import bg4 from '../assets/img/bg_carousel/bg4.jpg';
 import 'swiper/css';
 
 import { Autoplay } from 'swiper/modules';
