+++ src/pages/Home.tsx
@@ -1,12 +1,12 @@
 import '../css/home.css';
-import Hero from './components/Welcome';
-import About from './components/History';
-import Services from './components/Activities';
-import Portfolio from './components/Gallery';
-import Testimonials from './components/Committee';
-import Contact from './components/Contact';
-import Footer from './components/Footer';
-import NewHeader from './components/NewHeader';
+import Hero from '../components/Welcome';
+import About from '../components/History';
+import Services from '../components/Activities';
+import Portfolio from '../components/Gallery';
+import Testimonials from '../components/Committee';
+import Contact from '../components/Contact';
+import Footer from '../components/Footer';
+import NewHeader from '../components/NewHeader';
 
 
 function Home() {
